[OPENBUS-2475] [Java] Assistente não pode aceitar tempo de espera entre tentativas igual a ZERO - Merge das correções realizadas no branch 02_00_01 (r151104, r151119, r151120)
git-svn-id: https://subversion.tecgraf.puc-rio.br/engdist/openbus/sdk/java/trunk@151155 ae0415b3-e90b-0410-900d-d0be9363c56b

+++ demo/assistant/dedicatedclock/src/demo/DedicatedClockClient.java
@@ -32,7 +32,7 @@ public final class DedicatedClockClient {
   private static int port;
   private static String entity;
   private static String password;
-  private static int interval = 1;
+  private static float interval = 1.0f;
   private static int retries = 10;
 
   /**
@@ -78,7 +78,7 @@ public final class DedicatedClockClient {
     // - intervalo entre falhas
     if (args.length > 4) {
       try {
-        interval = Integer.parseInt(args[4]);
+        interval = Float.parseFloat(args[4]);
       }
       catch (NumberFormatException e) {
         System.out.println("Valor de [interval] deve ser um número");
