+++ demo/assistant/dedicatedclock/src/demo/DedicatedClockServer.java
@@ -28,7 +28,7 @@ public final class DedicatedClockServer {
   private static int port;
   private static String entity;
   private static OpenBusPrivateKey privateKey;
-  private static int interval = 1;
+  private static float interval = 1.0f;
 
   /**
    * Função principal.
@@ -78,7 +78,7 @@ public final class DedicatedClockServer {
     }
     if (args.length > 4) {
       try {
-        interval = Integer.parseInt(args[4]);
+        interval = Float.parseFloat(args[4]);
       }
       catch (NumberFormatException e) {
         System.out.println("Valor de [interval] deve ser um número");
